Fix #175
